requirements.txt
streamlit
pandas
plotly
SQLAlchemy
psycopg2-binary
fastapi
uvicorn[standard]
confluent-kafka
python-dotenv
lifelines
prophet